+++ includes/AssetManager.php
@@ -203,13 +203,14 @@ class AssetManager extends BaseClass {
 	 */
 	public function get_debug_info() {
 
-		$info  = 'PHP:         ' . PHP_VERSION . PHP_EOL;
-		$info .= 'WP:          ' . get_bloginfo( 'version' ) . PHP_EOL;
-		$info .= 'Plugin:      ' . $this->plugin()->name() . ':v' . $this->plugin()->version() . PHP_EOL;
-		$info .= 'DOMDocument: ' . ( class_exists( DOMDocument::class ) ? '✓' : '✕' ) . PHP_EOL;
-		$info .= 'DOMXPath:    ' . ( class_exists( DOMXPath::class ) ? '✓' : '✕' ) . PHP_EOL;
-
-		return $info;
+		return [
+			'PHP'              => PHP_VERSION,
+			'WP'               => get_bloginfo( 'version' ),
+			'Plugin'           => $this->plugin()->name() . ':v' . $this->plugin()->version(),
+			'Action Scheduler' => class_exists( 'ActionScheduler_Versions' ) ? '✓' : '✕',
+			'DOMDocument'      => class_exists( DOMDocument::class ) ? '✓' : '✕',
+			'DOMXPath'         => class_exists( DOMXPath::class ) ? '✓' : '✕',
+		];
 	}
 
 	/**