+++ languages/wptelegram-js-translations.php
@@ -34,14 +34,14 @@ $generated_i18n_strings = array(
 	// Reference: ../../packages/js/shared-ui/components/instructions.tsx:17
 	__( 'INSTRUCTIONS!', 'wptelegram' ),
 
-	// Reference: ../../packages/js/shared-ui/components/plugin-info/plugin-info-card.tsx:57
+	// Reference: ../../packages/js/shared-ui/components/plugin-info/plugin-info-card.tsx:59
 	/* translators: %s: plugin name */
 	__( 'Do you like %s?', 'wptelegram' ),
 
-	// Reference: ../../packages/js/shared-ui/components/plugin-info/plugin-info-card.tsx:69
+	// Reference: ../../packages/js/shared-ui/components/plugin-info/plugin-info-card.tsx:71
 	__( 'Write a review', 'wptelegram' ),
 
-	// Reference: ../../packages/js/shared-ui/components/plugin-info/plugin-info-card.tsx:85
+	// Reference: ../../packages/js/shared-ui/components/plugin-info/plugin-info-card.tsx:87
 	__( 'Need help?', 'wptelegram' ),
 
 	// Reference: ../../packages/js/shared-ui/components/variable-button.tsx:25
@@ -75,7 +75,7 @@ $generated_i18n_strings = array(
 	__( 'Please wait…', 'wptelegram' ),
 
 	// Reference: ../../packages/js/shared-ui/form/use-bot-token-test.tsx:67
-	// Reference: js/settings/ui/basics/tab.tsx:37
+	// Reference: js/settings/ui/basics/tab.tsx:30
 	__( 'Test Token', 'wptelegram' ),
 
 	// Reference: ../../packages/js/shared-ui/form/use-chat-with-test.tsx:93
@@ -96,17 +96,25 @@ $generated_i18n_strings = array(
 	// Reference: ../../packages/js/shared-ui/wptelegram/channels-field.tsx:73
 	__( 'Add channel', 'wptelegram' ),
 
-	// Reference: ../../packages/js/shared-ui/wptelegram/delay-in-posting.tsx:29
+	// Reference: ../../packages/js/shared-ui/wptelegram/delay-in-posting.tsx:36
 	__( 'The delay starts after the post gets published.', 'wptelegram' ),
 
-	// Reference: ../../packages/js/shared-ui/wptelegram/delay-in-posting.tsx:34
+	// Reference: ../../packages/js/shared-ui/wptelegram/delay-in-posting.tsx:42
+	/* translators: %s is the plugin name */
+	__( 'If you face any issues with delay in posting, please install the %s plugin.', 'wptelegram' ),
+
+	// Reference: ../../packages/js/shared-ui/wptelegram/delay-in-posting.tsx:64
+	__( 'Minute(s)', 'wptelegram' ),
+
+	// Reference: ../../packages/js/shared-ui/wptelegram/delay-in-posting.tsx:71
 	__( 'Warning', 'wptelegram' ),
 
-	// Reference: ../../packages/js/shared-ui/wptelegram/delay-in-posting.tsx:38
+	// Reference: ../../packages/js/shared-ui/wptelegram/delay-in-posting.tsx:75
 	__( 'WordPress cron should not be disabled!', 'wptelegram' ),
 
-	// Reference: ../../packages/js/shared-ui/wptelegram/delay-in-posting.tsx:46
-	__( 'Minute(s)', 'wptelegram' ),
+	// Reference: ../../packages/js/shared-ui/wptelegram/delay-in-posting.tsx:80
+	/* translators: %s is the plugin name */
+	__( 'Otherwise, you may install the %s plugin.', 'wptelegram' ),
 
 	// Reference: ../../packages/js/shared-ui/wptelegram/disable-notification.tsx:18
 	__( 'Send the messages silently.', 'wptelegram' ),
@@ -524,7 +532,7 @@ $generated_i18n_strings = array(
 	__( 'profile', 'wptelegram' ),
 
 	// Reference: ../../packages/js/utilities/fields.ts:35
-	// Reference: js/settings/services/fields.ts:147
+	// Reference: js/settings/services/fields.ts:148
 	__( 'Invalid %s', 'wptelegram' ),
 
 	// Reference: ../../packages/js/utilities/fields.ts:39
@@ -560,14 +568,14 @@ $generated_i18n_strings = array(
 	// Reference: js/p2tg-block-editor/SendToTelegram.tsx:36
 	__( 'Send to Telegram', 'wptelegram' ),
 
-	// Reference: js/settings/services/fields.ts:127
+	// Reference: js/settings/services/fields.ts:128
 	__( 'channel', 'wptelegram' ),
 
-	// Reference: js/settings/services/fields.ts:169
+	// Reference: js/settings/services/fields.ts:171
 	/* translators: %s: field label */
 	__( 'At least one %s is required.', 'wptelegram' ),
 
-	// Reference: js/settings/services/fields.ts:170
+	// Reference: js/settings/services/fields.ts:172
 	__( 'chat ID', 'wptelegram' ),
 
 	// Reference: js/settings/services/fields.ts:21
@@ -588,13 +596,13 @@ $generated_i18n_strings = array(
 	// Reference: js/settings/services/fields.ts:28
 	__( 'If Email goes to', 'wptelegram' ),
 
-	// Reference: js/settings/ui/Sidebar.tsx:20
+	// Reference: js/settings/ui/Sidebar.tsx:21
 	__( 'Integrate your WordPress website perfectly with Telegram. Send posts automatically to Telegram when published or updated, whether to a Telegram Channel, Group or private chat, with full control. Get your email notifications on Telegram.', 'wptelegram' ),
 
-	// Reference: js/settings/ui/Sidebar.tsx:23
+	// Reference: js/settings/ui/Sidebar.tsx:24
 	__( 'Join our public chat on Telegram', 'wptelegram' ),
 
-	// Reference: js/settings/ui/Sidebar.tsx:28
+	// Reference: js/settings/ui/Sidebar.tsx:29
 	__( 'Support', 'wptelegram' ),
 
 	// Reference: js/settings/ui/TabbedSections.tsx:20
@@ -648,7 +656,7 @@ $generated_i18n_strings = array(
 	// Reference: js/settings/ui/basics/instructions.tsx:61
 	__( 'Configure the activated modules.', 'wptelegram' ),
 
-	// Reference: js/settings/ui/basics/tab.tsx:36
+	// Reference: js/settings/ui/basics/tab.tsx:29
 	__( 'Use %s above to set automatically.', 'wptelegram' ),
 
 	// Reference: js/settings/ui/notify/tab.tsx:33